It's now possible to nav to ItemDetailScreen
I made it so pressing a friend's purchase brings you to the detail screen of the purchase.
Eventually, the picture of the item will be replaced with a small window where you can see the listing.
I also got rid of SafeAreaView, as this caused a lot of issues. I instead use a margin at the top of the homepage that's iOS exclusive.