Fixes to queries
Marketplace collection is now compared correctly with purchases to get user categories and return recommendations

Logging has been cleaned up so it's not so uggo

All recommendations return objects now instead of just names. This is so that the user can navigate to the correct marketplace page when navigating from the shop

Items of categories purchased by the user now show up as expected. We will have to find a way to weigh items if the user purchases items of multiple categories so that the recommended is diverse.

Items bought by friends that the user has not already bought show up as expected and as objects. We need to weigh items from friends somehow. Maybe we can sort it by the items that are bought the most amount of times by friends and rank from there

We still need to do the actual ranking